Converted to Maven Project

+++ src/main/java/com/fxperiments/NumUtils.java
@@ -2,7 +2,7 @@ package com.fxperiments;
 
 public class NumUtils {
 	
-	public static boolean isThirteen(int num){
+	public static boolean isThirteen(byte num){
 		return num == 13;
 	}
